+++ packages/core/__tests__/slot-recipe.test.ts
@@ -3,7 +3,7 @@ import { processSlotRecipe } from './fixture'
 
 describe('slot recipe ruleset', () => {
   test('should work', () => {
-    expect(processSlotRecipe('button', { size: 'sm' })).toMatchInlineSnapshot(`
+    expect(processSlotRecipe('button', { size: 'sm' }).toCss()).toMatchInlineSnapshot(`
       "@layer recipes.slots {
           @layer _base {
               .button__container {
@@ -25,6 +25,71 @@ describe('slot recipe ruleset', () => {
               font-size: 2rem
           }
       }"
+    `)
+  })
+
+  test('assigned recipe with default jsx from slots', () => {
+    expect(processSlotRecipe('button', { size: 'sm' }).getRecipe('button')).toMatchInlineSnapshot(`
+      {
+        "baseName": "button",
+        "config": {
+          "base": {
+            "container": {
+              "fontFamily": "mono",
+            },
+            "icon": {
+              "fontSize": "1.5rem",
+            },
+          },
+          "className": "button",
+          "slots": [
+            "container",
+            "icon",
+          ],
+          "variants": {
+            "size": {
+              "md": {
+                "container": {
+                  "fontSize": "3rem",
+                  "lineHeight": "1.2em",
+                },
+              },
+              "sm": {
+                "container": {
+                  "fontSize": "5rem",
+                  "lineHeight": "1em",
+                },
+                "icon": {
+                  "fontSize": "2rem",
+                },
+              },
+            },
+          },
+        },
+        "dashName": "button",
+        "jsx": [
+          "Button",
+          "Button.Container",
+          "Button.Icon",
+        ],
+        "jsxName": "Button",
+        "match": /\\^Button\\$\\|\\^Button\\.Container\\$\\|\\^Button\\.Icon\\$/,
+        "props": [
+          "size",
+        ],
+        "splitProps": [Function],
+        "type": "recipe",
+        "upperName": "Button",
+        "variantKeyMap": {
+          "size": [
+            "sm",
+            "md",
+          ],
+        },
+        "variantKeys": [
+          "size",
+        ],
+      }
     `)
   })
 })