+++ packages/generator/src/artifacts/qwik-jsx/jsx.string-literal.ts
@@ -7,23 +7,32 @@ export function generateQwikJsxStringLiteralFactory(ctx: Context) {
   return {
     js: outdent`
     import { h } from '@builder.io/qwik'
+    ${ctx.file.import('getDisplayName', './factory-helper')}
     ${ctx.file.import('css, cx', '../css/index')}
 
     function createStyledFn(Dynamic) {
       return function styledFn(template) {
-          const baseClassName = css(template)
-          const ${componentName} = function ${componentName}(props) {
-              const { as: Element = Dynamic, ...elementProps } = props
-              const classes = () => cx(baseClassName, elementProps.className)
+          const styles = css.raw(template)
+          
+          const ${componentName} = (props) => {
+            const { as: Element = Dynamic.__base__ || Dynamic, ...elementProps } = props
+            
+            function classes() {
+              return cx(css(Dynamic.__styles__, styles), elementProps.className)
+            }
 
-              return h(Element, {
-                  ...elementProps,
-                  className: classes(),
-              })
+            return h(Element, {
+              ...elementProps,
+              className: classes(),
+            })
           }
 
-          const name = (typeof Dynamic === 'string' ? Dynamic : Dynamic.displayName || Dynamic.name) || 'Component'
+          const name = getDisplayName(Dynamic)
+        
           ${componentName}.displayName = \`${factoryName}.\${name}\`
+          ${componentName}.__styles__ = styles
+          ${componentName}.__base__ = Dynamic
+  
           return ${componentName}
         }
     }