+++ packages/generator/src/artifacts/react-jsx/jsx.string-literal.ts
@@ -7,24 +7,33 @@ export function generateReactJsxStringLiteralFactory(ctx: Context) {
   return {
     js: outdent`
     import { createElement, forwardRef } from 'react'
+    ${ctx.file.import('getDisplayName', './factory-helper')}
     ${ctx.file.import('css, cx', '../css/index')}
 
     function createStyledFn(Dynamic) {
       return function styledFn(template) {
-        const baseClassName = css(template)
-        const ${componentName} = /* @__PURE__ */ forwardRef(function ${componentName}(props, ref) {
-            const { as: Element = Dynamic, ...elementProps } = props
-            const classes = () => cx(baseClassName, elementProps.className)
+        const styles = css.raw(template)
 
-            return createElement(Element, {
-                ref,
-                ...elementProps,
-                className: classes(),
-            })
+        const ${componentName} = /* @__PURE__ */ forwardRef(function ${componentName}(props, ref) {
+          const { as: Element = Dynamic.__base__ || Dynamic, ...elementProps } = props
+          
+          function classes() {
+            return cx(css(Dynamic.__styles__, styles), elementProps.className)
+          }
+
+          return createElement(Element, {
+              ref,
+              ...elementProps,
+              className: classes(),
+          })
         })
 
-        const name = (typeof Dynamic === 'string' ? Dynamic : Dynamic.displayName || Dynamic.name) || 'Component'
+        const name = getDisplayName(Dynamic)
+        
         ${componentName}.displayName = \`${factoryName}.\${name}\`
+        ${componentName}.__styles__ = styles
+        ${componentName}.__base__ = Dynamic
+        
         return ${componentName}
       }
     }