+++ extension/vsix-builder/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @pandacss/vsix-builder
 
+## 0.17.0
+
 ## 0.16.0
 
 ## 0.15.5